+++ src/main/java/com/epam/esm/mapper/impl/TagRowMapper.java
@@ -1,6 +1,8 @@
 package com.epam.esm.mapper.impl;
 
 import com.epam.esm.model.Tag;
+
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import org.springframework.jdbc.core.RowMapper;
@@ -11,7 +13,7 @@ public class TagRowMapper implements RowMapper<Tag> {
     @Override
     public Tag mapRow(ResultSet rs, int rowNum) throws SQLException {
         Tag tag = new Tag();
-        tag.setId(rs.getObject("id", Long.class));
+        tag.setId(rs.getObject("id", BigInteger.class));
         tag.setName(rs.getNString("name"));
         return tag;
     }
