+++ src/main/java/com/epam/esm/repository/extractor/TagResultSetExtractor.java
@@ -1,6 +1,8 @@
 package com.epam.esm.repository.extractor;
 
 import com.epam.esm.model.Tag;
+
+import java.math.BigInteger;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import org.springframework.dao.DataAccessException;
@@ -13,7 +15,7 @@ public class TagResultSetExtractor implements ResultSetExtractor<Tag> {
     public Tag extractData(ResultSet rs) throws SQLException, DataAccessException {
         Tag tag = new Tag();
         if (rs.next()) {
-            tag.setId(rs.getObject("id", Long.class));
+            tag.setId(rs.getObject("id", BigInteger.class));
             tag.setName(rs.getNString("name"));
         }
         return tag;
